Agentic_AI/requirements.txt
phi>=0.6.7
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
transformers
torch
huggingface-hub